+++ tests/test_documents.py
@@ -3,13 +3,13 @@ from mock import patch, Mock
 
 from django.db import models
 from django.utils.translation import ugettext_lazy as _
-
 from elasticsearch_dsl import GeoPoint
 
-from django_elasticsearch_dsl.documents import DocType
 from django_elasticsearch_dsl import fields
+from django_elasticsearch_dsl.documents import DocType
 from django_elasticsearch_dsl.exceptions import (ModelFieldNotMappedError,
                                                  RedeclaredFieldError)
+from tests import ES_MAJOR_VERSION
 
 
 class Car(models.Model):
@@ -35,7 +35,7 @@ class Manufacturer(models.Model):
 
 
 class CarDocument(DocType):
-    color = fields.StringField()
+    color = fields.TextField()
     type = fields.StringField()
 
     def prepare_color(self, instance):
@@ -46,6 +46,7 @@ class CarDocument(DocType):
         model = Car
         index = 'car_index'
         related_models = [Manufacturer]
+        doc_type = 'car_document'
 
 
 class DocTypeTestCase(TestCase):
@@ -74,6 +75,7 @@ class DocTypeTestCase(TestCase):
             class Meta:
                 model = Car
                 ignore_signals = True
+
         self.assertTrue(CarDocument2._doc_type.ignore_signals)
 
     def test_auto_refresh_added(self):
@@ -81,6 +83,7 @@ class DocTypeTestCase(TestCase):
             class Meta:
                 model = Car
                 auto_refresh = False
+
         self.assertFalse(CarDocument2._doc_type.auto_refresh)
 
     def test_queryset_pagination_added(self):
@@ -88,6 +91,7 @@ class DocTypeTestCase(TestCase):
             class Meta:
                 model = Car
                 queryset_pagination = 120
+
         self.assertIsNone(CarDocument._doc_type.queryset_pagination)
         self.assertEqual(CarDocument2._doc_type.queryset_pagination, 120)
 
@@ -115,7 +119,7 @@ class DocTypeTestCase(TestCase):
     def test_to_field(self):
         doc = DocType()
         nameField = doc.to_field('name', Car._meta.get_field('name'))
-        self.assertIsInstance(nameField, fields.StringField)
+        self.assertIsInstance(nameField, fields.TextField)
         self.assertEqual(nameField._path, ['name'])
 
     def test_to_field_with_unknown_field(self):
@@ -124,18 +128,20 @@ class DocTypeTestCase(TestCase):
             doc.to_field('manufacturer', Car._meta.get_field('manufacturer'))
 
     def test_mapping(self):
+        text_type = 'string' if ES_MAJOR_VERSION == 2 else 'text'
+
         self.assertEqual(
             CarDocument._doc_type.mapping.to_dict(), {
                 'car_document': {
                     'properties': {
                         'name': {
-                            'type': 'string'
+                            'type': text_type
                         },
                         'color': {
-                            'type': 'string'
+                            'type': text_type
                         },
                         'type': {
-                            'type': 'string'
+                            'type': text_type
                         },
                         'price': {
                             'type': 'double'
@@ -186,6 +192,7 @@ class DocTypeTestCase(TestCase):
         doc = CarDocument()
         car = Car(name="Type 57", price=5400000.0,
                   not_indexed="not_indexex", pk=51)
+
         doc.update(car)
 
         self.action_buffer().add_doc_actions.assert_called_with(
@@ -214,6 +221,7 @@ class DocTypeTestCase(TestCase):
                   not_indexed="not_indexex", pk=51)
         car2 = Car(name=_("Type 42"), price=50000.0,
                    not_indexed="not_indexex", pk=31)
+
         doc.update([car, car2], action='update')
 
         self.action_buffer().add_doc_actions.assert_called_with(
@@ -234,4 +242,19 @@ class DocTypeTestCase(TestCase):
         )
         self.action_buffer().execute.assert_called_with(
             doc.connection
+        )
+
+    def test_model_instance_iterable_update_with_pagination(self):
+        doc = CarDocument()
+        car1 = Car()
+        car2 = Car()
+        car3 = Car()
+
+        doc.update([car1, car2, car3])
+
+        self.action_buffer().add_doc_actions.assert_called_with(
+            doc, [car1, car2, car3], 'index',
+        )
+        self.action_buffer().execute.assert_called_with(
+            doc.connection, refresh=True
         )